BUG: Resolve ITK remote module CMake build errors
When ITKVkFFTBackend is built as part of ITK as a remote module
the ITKVkFFTBackend/test directory is not included directly as a project subdirectory
but rather as part of a secondary build step, leading to some VkFFT CMake variables not
propagating directly from the source root directory. Here the build errors are resolved
by adding VkFFT variables to the global cache and re-including header directories
at the subdirectory level.