ENH: Add convolution scale metric comparison notebook
VkFFT speedup is useful in convolution, but does not hold a universal
advantage over spatial convolution. It is useful to quantify where the
performance tradeoff between spatial and FFT convolution exists so that
the fastest procedure can be preemptively selected. This commit adds a
notebook with two experiments giving performance data on a range of
kernel and image sizes for Gaussian blurring along with visual
comparison of different metrics' suitability for distinguishing the best
procedure for a given input set.